requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Database & ORM
sqlalchemy
psycopg2-binary  # Để kết nối với PostgreSQL

# Validation & Schemas
pydantic
pydantic-settings

# Testing
pytest
httpx            # TestClient của FastAPI yêu cầu httpx

# Authentication (nếu bạn dùng JWT)
python-multipart
python-jose[cryptography]
passlib[bcrypt]

alembic==1.13.1